refactor: Moved action file
