+++ src/drive/web/modules/views/Drive/FileWithActions.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import useActions from './useActions'
+import useActions from 'drive/web/modules/actions/useActions'
 import { FileWithSelection } from 'drive/web/modules/filelist/File'
 
 const FileWithActions = props => {
